feat: enhanced documentation on account management edges
