+++ src/core/server/app/handlers/api/user/index.ts
@@ -1,4 +1,1 @@
 export * from "./download";
-export * from "./updateEmail";
-export * from "./updateUsername";
-export * from "./delete";
