+++ src/core/server/app/router/api/index.ts
@@ -56,7 +56,7 @@ export function createAPIRouter(app: AppOptions, options: RouterOptions) {
   // Create the auth router.
   router.use("/auth", createNewAuthRouter(app, options));
   router.use("/account", createNewAccountRouter(app, options));
-  router.use("/user", createNewUserRouter(app, options));
+  router.use("/user", createNewUserRouter(app));
 
   // Configure the GraphQL route.
   router.use(
